:sparkles: Adding signals finished and tested
*
