+++ Makefile
@@ -27,6 +27,7 @@ SRC_FILES	=	minishell.c \
 				lexer/lexer.c lexer/tokenize.c lexer/copy.c lexer/delimiter.c lexer/utils.c lexer/checker.c \
 				expanser/expanser.c expanser/var.c expanser/quote.c expanser/insert.c expanser/wildcards.c \
 				parser/parser.c parser/init.c parser/cmd.c parser/delete.c parser/redir.c parser/utils.c parser/path.c parser/file.c \
+				parser/heredoc/heredoc.c parser/heredoc/signals.c \
 				exec/execute.c exec/bracket.c exec/process.c exec/wait.c exec/fd.c \
 				utils/is.c utils/display.c utils/error.c utils/utils.c utils/ft_str_1.c utils/ft_str_2.c utils/split.c utils/get_next_line.c utils/itoa.c \
 				builtins/builtins.c builtins/ms_exit.c builtins/ms_env.c builtins/ms_echo.c builtins/ms_export.c builtins/ms_pwd.c builtins/ms_unset.c builtins/display.c \
