+++ src/expanser/expanser.c
@@ -6,7 +6,7 @@
 /*   By: owalsh <owalsh@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/24 13:13:33 by foctavia          #+#    #+#             */
-/*   Updated: 2022/09/20 12:26:08 by owalsh           ###   ########.fr       */
+/*   Updated: 2022/09/30 18:22:04 by owalsh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,21 +28,19 @@ static int	quote_expanser(t_token **tokens)
 	return (res);
 }
 
-static int	var_expanser(t_token **tokens, int exstatus)
+static int	var_expanser(t_token **tokens, int *res)
 {
-	int			res;
 	t_token		*tmp;
 
-	res = 0;
 	tmp = *tokens;
 	while (tmp && !g_global.data->err)
 	{
 		if (tmp->type == WORD && !ft_strncmp(tmp->value, "*", 1))
-			res = expanse_wildcard(&tmp);
+			*res = expanse_wildcard(&tmp);
 		if (tmp->type == VAR && !ft_strncmp(tmp->value, "$?", 2))
-			res = expanse_exstatus(&tmp, exstatus);
+			*res = expanse_exstatus(&tmp, *res);
 		if (tmp->type == VAR && tmp->var_stop > -1)
-			res = expanse_var(&tmp);
+			*res = expanse_var(&tmp, res);
 		if (!tmp)
 			break ;
 		tmp = tmp->next;
@@ -51,15 +49,15 @@ static int	var_expanser(t_token **tokens, int exstatus)
 	while (tmp)
 	{
 		if (tmp->type == VAR && tmp->var_stop > -1)
-			res = var_expanser(tokens, exstatus);
+			*res = var_expanser(tokens, res);
 		tmp = tmp->next;
 	}
-	return (res);
+	return (*res);
 }
 
-int	ms_expanser(t_token **tokens, int exstatus)
+int	ms_expanser(t_token **tokens, int *res)
 {
-	if (!var_expanser(tokens, exstatus))
+	if (!var_expanser(tokens, res))
 	{
 		if (!quote_expanser(tokens))
 		{
