+++ src/init/terminal.c
@@ -6,7 +6,7 @@
 /*   By: owalsh <owalsh@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/19 13:57:12 by owalsh            #+#    #+#             */
-/*   Updated: 2022/09/30 17:55:56 by owalsh           ###   ########.fr       */
+/*   Updated: 2022/09/30 18:42:11 by owalsh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	sig_exit(int signum)
 {
 	(void)signum;
-	write(2, "exit\n", 5);	
+	write(2, "exit\n", 5);
 	rl_clear_history();
 	clean(g_global.data);
 	exit(0);
@@ -26,6 +26,7 @@ void	sig_nl(int signum)
 	(void)signum;
 	if (rl_line_buffer[0])
 		return ;
+	write(1, "^C", 2);
 	write(0, "\n", 1);
 	rl_on_new_line();
 	rl_replace_line("", 0);
@@ -35,6 +36,6 @@ void	sig_nl(int signum)
 void	set_terminal(t_terminal *term)
 {
 	tcsetattr(STDIN_FILENO, TCSANOW, &term->new);
-    signal(SIGQUIT, &sig_exit);
-    signal(SIGINT, &sig_nl);
+	signal(SIGQUIT, &sig_exit);
+	signal(SIGINT, &sig_nl);
 }
