+++ src/lexer/checker.c
@@ -6,12 +6,11 @@
 /*   By: owalsh <owalsh@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/24 12:44:58 by owalsh            #+#    #+#             */
-/*   Updated: 2022/09/26 13:34:30 by owalsh           ###   ########.fr       */
+/*   Updated: 2022/09/30 18:11:49 by owalsh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
 
 static int	brk_order(t_token *head)
 {
